split http_events tests
